Move Incident Analysis placeholders to new framework
