--- patterns/test/test_incident_analysis/IncidentAnalysisTestSuiteDesktop.py
@@ -1,37 +1,0 @@
-# coding: utf-8
-# -----------------------------------------------------------------------------
-# Copyright 2015 Esri
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# -----------------------------------------------------------------------------
-
-# ==================================================
-# IncidentAnalysisTestSuiteDesktop.py
-# --------------------------------------------------
-# requirments: ArcGIS 10.X, Python 2.7
-# author: ArcGIS Solutions
-# company: Esri
-# ==================================================
-# description: <Description>
-# ==================================================
-# history:
-# <date> - <initals> - <modifications>
-# ==================================================
-
-import unittest
-import ClusterAnalysisTestCase
-import CountIncidentsByLOCTestCase
-import FindPercentChangeTestCase
-import HotSpotsByAreaTestCase
-import IncidentDensityTestCase
-import IncidentHotSpotsTestCase
-import IncidentTableToPointTestCase
