--- patterns/test/versions.xml
@@ -1,83 +1,0 @@
-<project name="ArcGIS Versions" default="set_all_variables" basedir=".">
-    <description>
-	OUTPUTS: ${dir.arcgisdesktop}, ${dir.arcpy}
-	Checks system environment variable and sets the above values depending on version detected.
-        Sets the correct ArcPy and ArcGIS Folder location depending on version detected.
-        IMPORTANT: If Python is not installed at the default location, 
-                   you will need to edit the task "SET_CUSTOM_PYTHON_LOCATION"
-    </description>
-
-<property name="default_python_location" value="C:\Python27" />
-
-<!-- ************************************************************************************************************ -->
-<!-- IMPORTANT: If you have *NOT* installed Python in the default install location, you will need to set the following: -->
-<!-- UNCOMMENT THIS PROPERTY: -->
-<!-- <property name="USE_CUSTOM_PYTHON_LOCATION" value="true"/> -->
-<!-- SET TO YOUR CUSTOM INSTALL LOCATION (the location with Python.exe): -->
-<target name="SET_CUSTOM_PYTHON_LOCATION" if="USE_CUSTOM_PYTHON_LOCATION">
-	<echo message="Setting Custom Values for ArcGIS Desktop and Python"/>
-	<property name="dir.arcpy" value="C:\MY_Python_XX\ArcGIS10.__SET_ME__"/>  <!-- CHANGE TO YOUR LOCATION -->
-</target>
-<!-- ************************************************************************************************************* -->
-
-<property environment="env"/>
-<echo message="AGSDESKTOPJAVA= ${env.AGSDESKTOPJAVA}"/>
-
-<!-- Set to the values for the current version/release (ex. 10.2): -->
-
-<property name="version.ags_current_version" value="10.3"/>
-<property name="dir.arcpy_current_version" value="${default_python_location}\ArcGIS${version.ags_current_version}"/>
-
-<!-- Set to the values for the previous version/release (ex. 10.1): -->
-
-<property name="version.ags_previous_version" value="10.2"/>
-<property name="dir.arcpy_previous_version" value="${default_python_location}\ArcGIS${version.ags_previous_version}"/>
-
-<target name="checkCurrentVersion">
-	<condition property="PresentCurrentVersion">
-		<available file="${dir.arcpy_current_version}\python.exe" type="file"/>
-	</condition>
-</target>
-
-<target name="checkPreviousVersion">
-	<condition property="PresentPreviousVersion">
-		<available file="${dir.arcpy_previous_version}\python.exe" type="file"/>
-	</condition>
-</target>
-
-<target name="setvariablesCurrentVersion" depends="checkCurrentVersion" if="PresentCurrentVersion">
-	<echo message="Current version of ArcGIS detected."/>
-	<property name="dir.arcpy" value="${dir.arcpy_current_version}"/>
-</target>
-
-<target name="setvariablesPreviousVersion" depends="checkPreviousVersion" if="PresentPreviousVersion">
-	<echo message="Previous version of ArcGIS detected."/>
-	<property name="dir.arcpy" value="${dir.arcpy_previous_version}"/>
-</target>
-
-<!-- Check that expected variables exists -->
-
-<target name="check_ags_exists" depends="SET_CUSTOM_PYTHON_LOCATION,setvariablesCurrentVersion,setvariablesPreviousVersion">
-	<property name="dir.arcgisdesktop" value="${env.AGSDESKTOPJAVA}" />
-	<condition property="agsnotfound">
-	  <or>
-		<not>
-			<isset property="env.AGSDESKTOPJAVA"/>
-		</not>
-    <not>
-		  	<available file="${env.AGSDESKTOPJAVA}" type="dir"/>
-		</not>      
-    <not>
-		  	<available file="${dir.arcpy}/python.exe" type="file"/>
-		</not>   
-	  </or>
-	</condition>
-	<fail message="**** COULD NOT FIND ArcGIS/ArcPy AT: ${dir.arcgisdesktop}, ${dir.arcpy} ****" if="agsnotfound"/>	
-</target>
-
-<target name="set_all_variables" depends="check_ags_exists">
-	<echo message="ArcGIS Desktop, ArcPy found at: ${dir.arcgisdesktop}, ${dir.arcpy}" />
-</target>
-	
-</project>
-
