+++ utils/test/patterns_tests/ClusterAnalysisTestCase.py
