+++ utils/test/patterns_tests/CountIncidentsByLOCTestCase.py
