+++ utils/test/patterns_tests/FindPercentChangeTestCase.py
