+++ utils/test/patterns_tests/HotSpotsByAreaTestCase.py
