+++ utils/test/patterns_tests/IncidentDensityTestCase.py
