+++ utils/test/patterns_tests/IncidentHotSpotsTestCase.py
